Avoid type format mismatch on Windows 64
On Windows 64 size() is a size_t which is 64 bit while %lo format
requires an unsigned long which is 32 bit.

+++ src/package.cpp
@@ -161,7 +161,7 @@ bool XVAPackage::Write(const std::string& file)
 		ptr += 8;
 		snprintf(ptr, 8, "%07d", 0);
 		ptr += 8;
-		snprintf(ptr, 12, "%011lo", data.size());
+		snprintf(ptr, 12, "%011lo", static_cast<long unsigned>(data.size()));
 		ptr += 12;
 		snprintf(ptr, 12, "%011d", 0);
 		ptr += 12;
